Fixed a bug with socket.io-client assuming connection to '/' nsp
socket.io-client assumed connection to '/' namespace by default - thus
ignoring attempts to join the '/' nsp.

This causes a problem because kiss.io doesn't demand a mounted '/'
namespace and does have no default namespaces.

This is a temporary fix, until a fix will be pulled to the
socket.io-client repo or will be fixed in a fork exclusively built for
kiss.io (kiss.io-client).